(native) Move <Link> to Link.tsx
