+++ packages/react-native/src/index.ts
@@ -1,7 +1,7 @@
-export { LinkProps } from "./HookLink";
+export { LinkProps } from "./Link";
 
 export * from "@curi/react-universal";
 
-import Link from "./HookLink";
+import Link from "./Link";
 
 export { Link };
