+++ packages/react-native/types/Link.d.ts
